Decoupled change-signature from vscode webview
